Fix: add shell=True to allow override of where command in the PATH
The command where appears to return an empty string under wine, which aborts the build.
This can be fixed using an emulated command based on shutil.which(). This fix works
provided that subprocess.call can detect the new command in PATH, which is only
possible using shell=True due to https://bugs.python.org/issue8557